rename transpile_c2eo to transpile
