+++ .github/workflows/build.yml
@@ -86,7 +86,7 @@ jobs:
       - name: Parsing
         run: |
           cd ./project/scripts
-          python3 transpile_c2eo.py ../tests/parsing -n
+          python3 transpile.py ../tests/parsing -n
 
   test:
     needs:
